# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: |
    ClinicFlow REST contract. Sessions resolve a database role. The HttpOnly, Secure,
    SameSite=Lax cookie contains an opaque native token by default for continuity,
    or an app-issued JWT when AUTH_SESSION_MODE=jwt is explicitly configured.
    Both modes are database-revocable with a fixed 12-hour lifetime, with no refresh
    endpoint. Changing modes invalidates old cookies. JWT_SIGNING_KEY must be a
    dedicated high-entropy secret of at least 32 UTF-8 bytes, configured before
    enabling JWT mode. Session issuance may return 503 AUTH_SESSION_UNCONFIGURED.
    Newly created passwords require at least 8 characters, ASCII letters and numbers,
    and at most 1024 UTF-8 bytes (enforced by the server).
    Super admins access all; clinic admins and receptionists are assignment-scoped;
    doctors access their own resources and relevant patients; patients access self only.
    All filters narrow authorized scope, never expand it. DELETE deactivates records
    with history. Lists contain real persisted data only. Queue clients poll every 30 seconds.
    Invalid input returns 400, unauthenticated 401, forbidden 403, missing 404,
    state/capacity conflicts 409, rate limiting 429, unavailable integrations 503.
servers:
  - url: /api
security:
  - NativeSession: []
paths:
...
  /public/guest-requests:
    post:
      operationId: createGuestRequest
      security: []
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/GuestRequestInput" } } } }
      responses:
        "201": { description: Immediately allocated appointment and private receipt (idempotent replay returns same ticket), content: { application/json: { schema: { $ref: "#/components/schemas/GuestReceipt" } } } }
  /public/guest-receipt:
    post:
      operationId: getGuestReceipt
      security: []
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/GuestReceiptInput" } } } }
      responses:
        "200": { description: Own request receipt only, content: { application/json: { schema: { $ref: "#/components/schemas/GuestReceipt" } } } }
  /guest-requests:
    get:
      operationId: listGuestRequests
      parameters:
        - { name: search, in: query, schema: { type: string, maxLength: 200 } }
        - { name: sort, in: query, schema: { type: string, enum: [createdAt, "-createdAt", fullName, "-fullName", date, "-date"], default: "-createdAt" } }
        - { $ref: "#/components/parameters/SessionId" }
        - { $ref: "#/components/parameters/StartTime" }
        - { name: clinicId, in: query, schema: { type: string } }
        - { name: branchId, in: query, schema: { type: string } }
        - { name: doctorId, in: query, schema: { type: string } }
        - { name: date, in: query, schema: { type: string, format: date } }
        - { name: status, in: query, schema: { type: string, enum: [pending, confirmed, rejected] } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, maximum: 100000 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
      responses:
        "200": { description: Scoped requests, content: { application/json: { schema: { $ref: "#/components/schemas/GuestRequestList" } } } }
  /guest-requests/{id}/decision:
    post:
      operationId: decideGuestRequest
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/GuestDecisionInput" } } } }
      responses:
        "200": { description: Decided request, content: { application/json: { schema: { $ref: "#/components/schemas/StaffGuestRequest" } } } }
...
  /public/qr/{reference}:
    get:
      operationId: resolveQr
      security: []
      parameters: [{ name: reference, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Active booking context without personal data, content: { application/json: { schema: { $ref: "#/components/schemas/QrContext" } } } }
        "404": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    NativeSession:
      type: apiKey
      in: cookie
      name: digiq_session
      description: Opaque native token (default) or explicitly enabled app-issued JWT. Both are database-revocable, fixed 12-hour HttpOnly/Secure/SameSite=Lax cookies. No refresh endpoint.
  responses:
    Error:
      description: Request failed
      content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
  parameters:
    SessionId: { name: sessionId, in: query, schema: { type: string } }
    StartTime: { name: startTime, in: query, schema: { type: string } }
...
  schemas:
...
    GuestRequestInput:
      type: object
      required: [qrReference, fullName, branchId, doctorId, date, requestId, receiptSecret]
      properties:
        sessionId: { type: string }
        startTime: { type: string }
        qrReference: { type: string, minLength: 1, maxLength: 200 }
        fullName: { type: string, minLength: 1, maxLength: 150 }
        email: { type: ["string", "null"], format: email, maxLength: 254 }
        mobile: { type: ["string", "null"], pattern: '^\+[1-9][0-9]{7,14}$', maxLength: 16 }
        branchId: { type: string, minLength: 1, maxLength: 100 }
        doctorId: { type: string, minLength: 1, maxLength: 100 }
        date: { type: string, pattern: '^\d{4}-\d{2}-\d{2}$' }
        requestId: { type: string, format: uuid }
        receiptSecret: { type: string, pattern: '^[a-f0-9]{64}$' }
    GuestReceiptInput:
      type: object
      required: [receiptSecret]
      properties:
        receiptSecret: { type: string, pattern: '^[a-f0-9]{64}$' }
    GuestDecisionInput:
      type: object
      required: [action, reason]
      properties:
        action: { type: string, enum: [confirm, reject] }
        reason: { type: string, minLength: 1, maxLength: 500 }
    GuestReceipt:
      type: object
      required: [id, status, fullName, clinicName, branchName, doctorName, date, startTime, endTime, timezone, token, reason, appointmentId, reference, branchAddress, appointmentStatus, revision, checkInUrl]
      properties:
        confirmationEmail: { $ref: "#/components/schemas/ConfirmationEmailOutcome" }
        dateFormat: { $ref: "#/components/schemas/ClinicDateFormat" }
        timeFormat: { $ref: "#/components/schemas/ClinicTimeFormat" }
        sessionId: { type: [string, "null"] }
        id: { type: string }
        status: { type: string, enum: [pending, confirmed, rejected] }
        fullName: { type: string }
        clinicName: { type: string }
        branchName: { type: string }
        doctorName: { type: string }
        date: { type: string }
        startTime: { type: ["string", "null"] }
        endTime: { type: ["string", "null"] }
        timezone: { type: string }
        token: { type: ["string", "null"] }
        reason: { type: ["string", "null"] }
        appointmentId: { type: ["string", "null"] }
        reference: { type: ["string", "null"] }
        branchAddress: { type: ["string", "null"] }
        appointmentStatus: { type: ["string", "null"] }
        revision: { type: ["integer", "null"] }
        checkInUrl: { type: ["string", "null"], description: "Personal signed check-in QR URL. Keep private like receiptSecret; only returned via booking or receipt capability." }
    StaffGuestRequest:
      allOf:
        - { $ref: "#/components/schemas/GuestReceipt" }
        - type: object
          required: [clinicId, branchId, doctorId, email, mobile, appointmentId, createdAt]
          properties:
            clinicId: { type: string }
            branchId: { type: string }
            doctorId: { type: string }
            email: { type: ["string", "null"] }
            mobile: { type: ["string", "null"] }
            appointmentId: { type: ["string", "null"] }
            createdAt: { type: string, format: date-time }
    GuestRequestList:
      type: object
      required: [items, total, page, pageSize]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/StaffGuestRequest" } }
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
...
    ApiError:
      type: object
      required: [error]
      properties:
        error: { type: string }
        code: { type: string }
...
    ClinicDateFormat:
      type: string
      enum: ["DD MMM YYYY", "DD/MM/YYYY", "MM/DD/YYYY", "YYYY-MM-DD"]
      description: Parent Clinic Group display preference; locations inherit it. Default DD MMM YYYY.
    ClinicTimeFormat:
      type: string
      enum: ["12h", "24h"]
      description: Parent Clinic Group display preference; locations inherit it. Default 12h.
...
    ConfirmationEmailOutcome:
      type: string
      enum: [provider_accepted, unavailable, disabled, no_recipient, not_attempted]
      description: Booking confirmation email attempt outcome. Provider acceptance is not proof of inbox delivery.
...
    QrContext:
      type: object
      required: [reference, clinicId, clinicName]
      properties:
        dateFormat: { $ref: "#/components/schemas/ClinicDateFormat" }
        timeFormat: { $ref: "#/components/schemas/ClinicTimeFormat" }
        reference: { type: string }
        clinicId: { type: string }
        clinicName: { type: string }
        clinicAddress: { type: [string, "null"] }
        branchAddress: { type: [string, "null"] }
        branchCity: { type: [string, "null"] }
        branchTimezone: { type: [string, "null"] }
        branchId: { type: [string, "null"] }
        branchName: { type: [string, "null"] }
        doctorId: { type: [string, "null"] }
        doctorName: { type: [string, "null"] }
...